@@ -19,14 +19,14 @@
         "resolve-url-loader": "^5.0.0",
         "sass": "^1.32.11",
         "sass-loader": "^11.0.1",
-        "vue": "^2.6.12",
-        "vue-loader": "^15.9.7",
-        "vue-template-compiler": "^2.6.12"
+        "vue-loader": "^16.8.3"
     },
     "dependencies": {
         "@toast-ui/editor": "^3.1.5",
         "autoprefixer": "^10.4.5",
         "codemirror": "^5.65.3",
-        "jquery": "^3.6.1"
+        "jquery": "^3.6.1",
+        "mitt": "^3.0.0",
+        "vue": "^3.2.45"
     }
 }
