@@ -1,17 +1,18 @@
-window.Vue = require('vue').default;
-window.axios = require('axios');
-window.axios.defaults.headers.common['X-Requested-With'] = 'XMLHttpRequest';
+import { createApp } from 'vue';
+import mitt from 'mitt';
 
-/**
- * Register Vue components here.
- */
-Vue.component('like', () => import('../js/Components/Like.vue'));
-Vue.component('follow', () => import('../js/Components/Follow.vue'));
-Vue.component('followcounter', () => import('../js/Components/FollowCounter.vue'));
-Vue.component('likecounter', () => import('../js/Components/LikeCounter.vue'));
+import Follow  from './Components/Follow.vue';
+import FollowCounter  from './Components/FollowCounter.vue';
+import Like from './Components/Like.vue';
+import LikeCounter from './Components/LikeCounter.vue';
 
-export const EventBus = new Vue();
+const emitter = mitt();
+const app = createApp({});
+app.config.globalProperties.emitter = emitter
 
-const app = new Vue({
-    el:'#app'
-});
+app.component('like', Like);
+app.component('likecounter', LikeCounter);
+app.component('follow', Follow);
+app.component('followcounter', FollowCounter);
+
+app.mount('#app');
